three wisemen

+++ setup.go
@@ -104,7 +104,7 @@ func generateQuotesTable(path string) error {
 		}
 		id := uuid.New().ID()
 
-		params := database.CreateQuoteParams{ID: id, Quote: quote, Author: author, Categories: "default"}
+		params := database.CreateQuoteParams{ID: int64(id), Quote: quote, Author: author, Categories: "default"}
 		_, err := queries.CreateQuote(ctx, params)
 		if err != nil {
 			return err
